NewsAnalyser/requirements.txt
matplotlib
pandas
streamlit
openpyxl
nltk
scikit-learn
textblob
plotly
boto3
botocore
azure-storage-blob
google-cloud-storage
bertopic
texthero
spacy
numpy
six
pyldavis